+++ scripts/__init__.py
@@ -1,0 +1,3 @@
+"""
+CFE Tariff Analyzer - Scripts Package
+"""